pyproject.toml
[project]
name = "rag-qdrant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "fastembed>=0.8.0",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.17",
    "langchain-classic>=1.0.7",
    "langchain-cohere>=0.5.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.3",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.2.1",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.1.10",
    "langserve>=0.3.0",
    "pymupdf>=1.27.2.3",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "ragas>=0.4.3",
    "sse-starlette>=2.1.0",
    "unstructured-inference>=1.6.11",
    "unstructured[pdf]>=0.22.27",
    "uv>=0.11.11",
    "uvicorn[standard]>=0.32.0",
]
